Added hack that converts bondorders for specific groups to 1.5.
The following groups are affected:
- all aromatic bonds get bond order 1.5. This information is taken
  from OpenBabel.
- cm-om bonds in COO-
- no-on bonds in NO2
- s4-om bonds in SO4--
- p5-om bonds in PO4---
The latter four changes are hard-coded so if we change names for
atom types the code will have to be updated.

+++ src/alexandria/mymol.cpp
@@ -1098,7 +1098,7 @@ void MyMol::addShells(const Poldata *pd,
     bHaveShells_ = true;
 }
 
-int MyMol::bondOrder(int ai, int aj) const
+double MyMol::bondOrder(int ai, int aj) const
 {
     if (bondOrder_.find({ai, aj}) != bondOrder_.end())
     {
